Deduplicate Call definition
